+++ examples/kitchen-sink/shared/template-engine.js
@@ -1,5 +1,14 @@
 import Kensington from 'kensington';
 
+export {
+  computed,
+  effect,
+  isBrowser,
+  registerComponents,
+  renderForHydration,
+  signal,
+} from 'kensington';
+
 class TaskListEngine extends Kensington {
   sortableList = this.createCustomTag('k-sortable-list', { onreorder: Function });
 }